refactor(core): Move leaf entities to `@n8n/db`
