+++ packages/@n8n/db/src/entities/index.ts
@@ -1,0 +1,19 @@
+import { EventDestinations } from './event-destinations';
+import { InstalledNodes } from './installed-nodes';
+import { InstalledPackages } from './installed-packages';
+import { InvalidAuthToken } from './invalid-auth-token';
+import { ProcessedData } from './processed-data';
+import { Settings } from './settings';
+import { Variables } from './variables';
+import { WebhookEntity } from './webhook-entity';
+
+export {
+	EventDestinations,
+	InstalledNodes,
+	InstalledPackages,
+	InvalidAuthToken,
+	ProcessedData,
+	Settings,
+	Variables,
+	WebhookEntity,
+};
