+++ packages/@n8n/db/src/entities/installed-nodes.ts
