+++ packages/@n8n/db/src/entities/installed-packages.ts
@@ -1,6 +1,6 @@
-import { WithTimestamps } from '@n8n/db';
 import { Column, Entity, JoinColumn, OneToMany, PrimaryColumn } from '@n8n/typeorm';
 
+import { WithTimestamps } from './abstract-entity';
 import type { InstalledNodes } from './installed-nodes';
 
 @Entity()
