+++ packages/@n8n/db/src/entities/processed-data.ts
@@ -1,8 +1,7 @@
-import { JsonColumn, WithTimestamps } from '@n8n/db';
 import { Entity, PrimaryColumn } from '@n8n/typeorm';
+import type { IProcessedDataEntries, IProcessedDataLatest } from 'n8n-workflow';
 
-import type { IProcessedDataEntries, IProcessedDataLatest } from '@/types-db';
-
+import { JsonColumn, WithTimestamps } from './abstract-entity';
 import { objectRetriever } from '../utils/transformers';
 
 @Entity()
