+++ packages/@n8n/db/src/index.ts
@@ -12,3 +12,15 @@ export {
 export { generateNanoId } from './utils/generators';
 export { isStringArray } from './utils/is-string-array';
 export { separate } from './utils/separate';
+export { idStringifier, lowerCaser, objectRetriever, sqlite } from './utils/transformers';
+
+export {
+	EventDestinations,
+	InstalledNodes,
+	InstalledPackages,
+	InvalidAuthToken,
+	ProcessedData,
+	Settings,
+	Variables,
+	WebhookEntity,
+} from './entities';
