+++ packages/cli/src/commands/__tests__/community-node.test.ts
@@ -1,8 +1,8 @@
+import { type InstalledNodes } from '@n8n/db';
 import { type Config } from '@oclif/core';
 import { mock } from 'jest-mock-extended';
 
 import { type CredentialsEntity } from '@/databases/entities/credentials-entity';
-import { type InstalledNodes } from '@/databases/entities/installed-nodes';
 import { type User } from '@/databases/entities/user';
 
 import { CommunityNode } from '../community-node';
