+++ packages/cli/src/controllers/community-packages.controller.ts
@@ -1,3 +1,4 @@
+import type { InstalledPackages } from '@n8n/db';
 import { Delete, Get, Patch, Post, RestController, GlobalScope } from '@n8n/decorators';
 
 import {
@@ -5,7 +6,6 @@ import {
 	STARTER_TEMPLATE_NAME,
 	UNKNOWN_FAILURE_REASON,
 } from '@/constants';
-import type { InstalledPackages } from '@/databases/entities/installed-packages';
 import { BadRequestError } from '@/errors/response-errors/bad-request.error';
 import { InternalServerError } from '@/errors/response-errors/internal-server.error';
 import { EventService } from '@/events/event.service';
