+++ packages/cli/src/databases/entities/execution-entity.ts
@@ -1,4 +1,4 @@
-import { DateTimeColumn, datetimeColumnType } from '@n8n/db';
+import { DateTimeColumn, datetimeColumnType, idStringifier } from '@n8n/db';
 import {
 	Column,
 	Entity,
@@ -19,7 +19,6 @@ import type { ExecutionAnnotation } from '@/databases/entities/execution-annotat
 import type { ExecutionData } from './execution-data';
 import type { ExecutionMetadata } from './execution-metadata';
 import { WorkflowEntity } from './workflow-entity';
-import { idStringifier } from '../utils/transformers';
 
 @Entity()
 @Index(['workflowId', 'id'])
