+++ packages/cli/src/databases/entities/index.ts
@@ -1,3 +1,14 @@
+import {
+	EventDestinations,
+	InstalledNodes,
+	InstalledPackages,
+	InvalidAuthToken,
+	ProcessedData,
+	Settings,
+	Variables,
+	WebhookEntity,
+} from '@n8n/db';
+
 import { AnnotationTagEntity } from './annotation-tag-entity.ee';
 import { AnnotationTagMapping } from './annotation-tag-mapping.ee';
 import { ApiKey } from './api-key';
@@ -5,20 +16,14 @@ import { AuthIdentity } from './auth-identity';
 import { AuthProviderSyncHistory } from './auth-provider-sync-history';
 import { AuthUser } from './auth-user';
 import { CredentialsEntity } from './credentials-entity';
-import { EventDestinations } from './event-destinations';
 import { ExecutionAnnotation } from './execution-annotation.ee';
 import { ExecutionData } from './execution-data';
 import { ExecutionEntity } from './execution-entity';
 import { ExecutionMetadata } from './execution-metadata';
 import { Folder } from './folder';
 import { FolderTagMapping } from './folder-tag-mapping';
-import { InstalledNodes } from './installed-nodes';
-import { InstalledPackages } from './installed-packages';
-import { InvalidAuthToken } from './invalid-auth-token';
-import { ProcessedData } from './processed-data';
 import { Project } from './project';
 import { ProjectRelation } from './project-relation';
-import { Settings } from './settings';
 import { SharedCredentials } from './shared-credentials';
 import { SharedWorkflow } from './shared-workflow';
 import { TagEntity } from './tag-entity';
@@ -27,8 +32,6 @@ import { TestDefinition } from './test-definition.ee';
 import { TestMetric } from './test-metric.ee';
 import { TestRun } from './test-run.ee';
 import { User } from './user';
-import { Variables } from './variables';
-import { WebhookEntity } from './webhook-entity';
 import { WorkflowEntity } from './workflow-entity';
 import { WorkflowHistory } from './workflow-history';
 import { WorkflowStatistics } from './workflow-statistics';
