+++ packages/cli/src/databases/repositories/event-destinations.repository.ts
@@ -1,7 +1,6 @@
+import { EventDestinations } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
-
-import { EventDestinations } from '../entities/event-destinations';
 
 @Service()
 export class EventDestinationsRepository extends Repository<EventDestinations> {
