+++ packages/cli/src/databases/repositories/installed-nodes.repository.ts
@@ -1,7 +1,6 @@
+import { InstalledNodes } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
-
-import { InstalledNodes } from '../entities/installed-nodes';
 
 @Service()
 export class InstalledNodesRepository extends Repository<InstalledNodes> {
