+++ packages/cli/src/databases/repositories/installed-packages.repository.ts
@@ -1,9 +1,9 @@
+import { InstalledPackages } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
 import type { PackageDirectoryLoader } from 'n8n-core';
 
 import { InstalledNodesRepository } from './installed-nodes.repository';
-import { InstalledPackages } from '../entities/installed-packages';
 
 @Service()
 export class InstalledPackagesRepository extends Repository<InstalledPackages> {
