+++ packages/cli/src/databases/repositories/invalid-auth-token.repository.ts
@@ -1,7 +1,6 @@
+import { InvalidAuthToken } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
-
-import { InvalidAuthToken } from '../entities/invalid-auth-token';
 
 @Service()
 export class InvalidAuthTokenRepository extends Repository<InvalidAuthToken> {
