+++ packages/cli/src/databases/repositories/processed-data.repository.ts
@@ -1,7 +1,6 @@
+import { ProcessedData } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
-
-import { ProcessedData } from '../entities/processed-data';
 
 @Service()
 export class ProcessedDataRepository extends Repository<ProcessedData> {
