+++ packages/cli/src/databases/repositories/settings.repository.ts
@@ -1,11 +1,10 @@
+import { Settings } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
 import { ErrorReporter } from 'n8n-core';
 
 import config from '@/config';
 import { EXTERNAL_SECRETS_DB_KEY } from '@/external-secrets.ee/constants';
-
-import { Settings } from '../entities/settings';
 
 @Service()
 export class SettingsRepository extends Repository<Settings> {
