+++ packages/cli/src/databases/repositories/variables.repository.ts
@@ -1,7 +1,6 @@
+import { Variables } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
-
-import { Variables } from '../entities/variables';
 
 @Service()
 export class VariablesRepository extends Repository<Variables> {
