+++ packages/cli/src/databases/repositories/webhook.repository.ts
@@ -1,7 +1,6 @@
+import { WebhookEntity } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, Repository } from '@n8n/typeorm';
-
-import { WebhookEntity } from '../entities/webhook-entity';
 
 @Service()
 export class WebhookRepository extends Repository<WebhookEntity> {
