+++ packages/cli/src/environments.ee/source-control/__tests__/source-control.service.test.ts
@@ -1,4 +1,5 @@
 import type { SourceControlledFile } from '@n8n/api-types';
+import type { Variables } from '@n8n/db';
 import { Container } from '@n8n/di';
 import { mock } from 'jest-mock-extended';
 import { InstanceSettings } from 'n8n-core';
@@ -6,7 +7,6 @@ import { InstanceSettings } from 'n8n-core';
 import type { FolderWithWorkflowAndSubFolderCount } from '@/databases/entities/folder';
 import type { TagEntity } from '@/databases/entities/tag-entity';
 import type { User } from '@/databases/entities/user';
-import type { Variables } from '@/databases/entities/variables';
 import type { FolderRepository } from '@/databases/repositories/folder.repository';
 import type { TagRepository } from '@/databases/repositories/tag.repository';
 import { SourceControlPreferencesService } from '@/environments.ee/source-control/source-control-preferences.service.ee';
