+++ packages/cli/src/environments.ee/variables/variables.service.ee.ts
@@ -1,7 +1,7 @@
 import { generateNanoId } from '@n8n/db';
+import type { Variables } from '@n8n/db';
 import { Service } from '@n8n/di';
 
-import type { Variables } from '@/databases/entities/variables';
 import { VariablesRepository } from '@/databases/repositories/variables.repository';
 import { VariableCountLimitReachedError } from '@/errors/variable-count-limit-reached.error';
 import { VariableValidationError } from '@/errors/variable-validation.error';
