+++ packages/cli/src/eventbus/message-event-bus-destination/message-event-bus-destination-from-db.ts
@@ -1,8 +1,7 @@
+import type { EventDestinations } from '@n8n/db';
 import { Container } from '@n8n/di';
 import { Logger } from 'n8n-core';
 import { MessageEventBusDestinationTypeNames } from 'n8n-workflow';
-
-import type { EventDestinations } from '@/databases/entities/event-destinations';
 
 import { MessageEventBusDestinationSentry } from './message-event-bus-destination-sentry.ee';
 import { MessageEventBusDestinationSyslog } from './message-event-bus-destination-syslog.ee';
