+++ packages/cli/src/services/hooks.service.ts
@@ -1,3 +1,4 @@
+import type { Settings } from '@n8n/db';
 import { Service } from '@n8n/di';
 // eslint-disable-next-line n8n-local-rules/misplaced-n8n-typeorm-import
 import type { FindManyOptions, FindOneOptions, FindOptionsWhere } from '@n8n/typeorm';
@@ -8,7 +9,6 @@ import type { NextFunction, Response } from 'express';
 import { AuthService } from '@/auth/auth.service';
 import type { AuthUser } from '@/databases/entities/auth-user';
 import type { CredentialsEntity } from '@/databases/entities/credentials-entity';
-import type { Settings } from '@/databases/entities/settings';
 import type { User } from '@/databases/entities/user';
 import type { WorkflowEntity } from '@/databases/entities/workflow-entity';
 import { AuthUserRepository } from '@/databases/repositories/auth-user.repository';
