+++ packages/cli/src/webhooks/__tests__/webhook.service.test.ts
@@ -1,10 +1,10 @@
+import { WebhookEntity } from '@n8n/db';
 import { mock } from 'jest-mock-extended';
 import type { INode, INodeType, IWebhookData, IWorkflowExecuteAdditionalData } from 'n8n-workflow';
 import { Workflow } from 'n8n-workflow';
 import { v4 as uuid } from 'uuid';
 
 import config from '@/config';
-import { WebhookEntity } from '@/databases/entities/webhook-entity';
 import type { WebhookRepository } from '@/databases/repositories/webhook.repository';
 import type { NodeTypes } from '@/node-types';
 import type { CacheService } from '@/services/cache/cache.service';
