+++ packages/cli/test/integration/variables.test.ts
@@ -1,7 +1,7 @@
 import { generateNanoId } from '@n8n/db';
+import type { Variables } from '@n8n/db';
 import { Container } from '@n8n/di';
 
-import type { Variables } from '@/databases/entities/variables';
 import { VariablesRepository } from '@/databases/repositories/variables.repository';
 import { VariablesService } from '@/environments.ee/variables/variables.service.ee';
 import { CacheService } from '@/services/cache/cache.service';
